(front) update paths to renamed files
