+++ frontend/src/app/document-editor/document-editor.component.html
@@ -4,7 +4,7 @@
     </nb-tab>
     <nb-tab tabTitle="Sellos" *ngIf="documentId" class="stamps-tab">
         <ng-template nbTabContent>
-            <app-document-signatures></app-document-signatures>
+            <app-document-stamps></app-document-stamps>
         </ng-template>
     </nb-tab>
     <nb-tab tabTitle="Acceso compartido" *ngIf="documentId" class="signatures-tab">
