+++ frontend/src/app/document-editor/document-stamps/document-stamps.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { StampSelectorComponent } from './signature-selector/signature-selector.component';
-import { DocumentStampsComponent } from './document-signatures.component';
+import { StampSelectorComponent } from './stamp-selector/stamp-selector.component';
+import { DocumentStampsComponent } from './document-stamps.component';
 import { NbButtonModule, NbCardModule, NbContextMenuModule, NbDialogModule, NbIconModule, NbInputModule, NbListModule, NbSelectModule, NbSpinnerModule, NbTooltipModule } from '@nebular/theme';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { SharedModule } from '../../shared/shared.module';
